add services tests and fix others
